kernel: get rid of str_len global
We do this by changing ddG_PolyInput2Matrix to not copy its three
string arguments, but rather modify them directly -- which is fine
in our case, as in the one place calling it, the arguments are in
turn copies from another source.

Also remove the RATPLIST_STR and PLIST_STR helpers, they were just
aliases for CSTR_STRING (and we now use CONST_CSTR_STRING instead).