Replace unsafe PLIST APIs by safe ASS_LIST
Using ASS_LIST automatically sets the list size correctly, and also calls
CHANGED_BAG as needed. In the old code, CHANGED_BAG was missing from some
places it was needed (which could potentially result in a crash); and also was
called in some places it was *not* needed.

The resulting code is shorter and easier maintain, at the cost of a few
microseconds, which should not matter in practice, because the expensive
operations happen inside Cdd.